app/bot/requirements.txt
aiogram==2.27
redis>=4.5
cryptography>=41
python-dotenv>=1.0
requests>=2.31